% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{discharge_coef}
\alias{discharge_coef}
\title{Discharge coefficient}
\usage{
discharge_coef(Q, D, rpm)
}
\arguments{
\item{Q}{Turbine discharge (cfs)}

\item{D}{Nominal diameter (ft) of runner}

\item{rpm}{Runner revolutions per minute}
}
\description{
Non-dimensional discharge coefficient term incorporated into strike equations
}
