% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{energy_coef}
\alias{energy_coef}
\title{Energy coefficient}
\usage{
energy_coef(H, D, rpm)
}
\arguments{
\item{H}{Net head on the turbine (ft)}

\item{D}{Nominal diameter (ft) of runner}

\item{rpm}{Runner revolutions per minute}
}
\description{
Non-dimensional energy coefficient term incorporated into strike equations
}
