% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/francis.R
\name{francis_alpha}
\alias{francis_alpha}
\title{Francis alpha}
\usage{
francis_alpha(Q, H, D, D1, D2, rpm, eta, opt, xi, B)
}
\arguments{
\item{Q}{Turbine discharge (cfs)}

\item{H}{Net head on the turbine (ft)}

\item{D}{Nominal diameter (ft) of runner}

\item{D1}{Diameter (ft) at the intake of the runner}

\item{D2}{Diameter (ft) at the outlet of the runner}

\item{rpm}{Runner revolutions per minute}

\item{eta}{Turbine efficiency}

\item{opt}{Ratio of turbine discharge at best efficiency to hydraulic capacity}

\item{xi}{Ratio between Q with no exit swirl and Qopt}

\item{B}{Runner height (ft) at inlet}
}
\description{
Angle (rad) to tangential of absolute flow upstream of runner
}
