% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/francis.R
\name{francis_beta}
\alias{francis_beta}
\title{Francis beta}
\usage{
francis_beta(Q, D, D1, D2, rpm, opt, xi)
}
\arguments{
\item{Q}{Turbine discharge (cfs)}

\item{D}{Nominal diameter (ft) of runner}

\item{D1}{Diameter (ft) at the intake of the runner}

\item{D2}{Diameter (ft) at the outlet of the runner}

\item{rpm}{Runner revolutions per minute}

\item{opt}{Ratio of turbine discharge at best efficiency to hydraulic capacity}

\item{xi}{Ratio between Q with no exit swirl and Qopt}
}
\description{
Relative flow angle (rad) at turbine discharge; used in strike equations
}
