% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/francis.R
\name{francis_strike}
\alias{francis_strike}
\title{Francis blade strike probability}
\usage{
francis_strike(Q, H, D, D1, D2, rpm, eta, opt, xi, B, N, L, lambda = 0.2)
}
\arguments{
\item{Q}{Turbine discharge (cfs)}

\item{H}{Net head on the turbine (ft)}

\item{D}{Nominal diameter (ft) of runner}

\item{D1}{Diameter (ft) at the intake of the runner}

\item{D2}{Diameter (ft) at the outlet of the runner}

\item{rpm}{Runner revolutions per minute}

\item{eta}{Turbine efficiency}

\item{opt}{Ratio of turbine discharge at best efficiency to hydraulic capacity}

\item{xi}{Ratio between Q with no exit swirl and Qopt}

\item{B}{Runner height (ft) at inlet}

\item{N}{Number of blades}

\item{L}{Fish length (ft)}

\item{lambda}{Actual mortality correlation; influenced by many factors including unit type and fish species}
}
\description{
Calculates leading-edge blade strike probability from a Francis turbine
}
