% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propeller.R
\name{propeller_alpha}
\alias{propeller_alpha}
\title{Propeller alpha}
\usage{
propeller_alpha(Q, H, D, rpm, eta, opt, r)
}
\arguments{
\item{Q}{Turbine discharge (cfs)}

\item{H}{Net head on the turbine (ft)}

\item{D}{Nominal diameter (ft) of runner}

\item{rpm}{Runner revolutions per minute}

\item{eta}{Turbine efficiency}

\item{opt}{Ratio of turbine discharge at best efficiency to hydraulic capacity}

\item{r}{Radius at mid blade (ft)}
}
\description{
Angle (rad) to tangential of absolute flow upstream of runner
}
