% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propeller.R
\name{propeller_beta}
\alias{propeller_beta}
\title{Propeller beta}
\usage{
propeller_beta(Q, D, rpm, opt, r)
}
\arguments{
\item{Q}{Turbine discharge (cfs)}

\item{D}{Nominal diameter (ft) of runner}

\item{rpm}{Runner revolutions per minute}

\item{opt}{Ratio of turbine discharge at best efficiency to hydraulic capacity}

\item{r}{Radius at mid blade (ft)}
}
\description{
Relative flow angle (rad) at turbine discharge; used in strike equations
}
