% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbsa.R
\name{tbsa}
\alias{tbsa}
\title{Turbine blade strike analysis}
\usage{
tbsa(fish_num, length_mean, length_sd, route_data)
}
\arguments{
\item{fish_num}{Number of fish in the simulation}

\item{length_mean}{Mean fish length (ft)}

\item{length_sd}{Standard deviation of fish length (ft)}

\item{route_data}{Data frame with input parameters for each route (see route_data_ex for example)}
}
\description{
Runs stochastic simulation of turbine blade strike analysis based on input parameters in route_data
}
\examples{
tbsa(10, 1.5, 0.25, route_data_ex)
}
