% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan.R
\name{kaplan_alpha}
\alias{kaplan_alpha}
\title{Kaplan alpha}
\usage{
kaplan_alpha(Q, H, D, rpm, eta, r)
}
\arguments{
\item{Q}{Turbine discharge (cfs)}

\item{H}{Net head on the turbine (ft)}

\item{D}{Nominal diameter (ft) of runner}

\item{rpm}{Runner revolutions per minute}

\item{eta}{Turbine efficiency}

\item{r}{Radius at mid blade (ft)}
}
\description{
Angle (rad) to tangential of absolute flow upstream of runner
}
