% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propeller.R
\name{propeller_beta}
\alias{propeller_beta}
\title{Propeller beta}
\usage{
propeller_beta(Q, D, rpm, opt, radius_ratio)
}
\arguments{
\item{Q}{Turbine discharge (cfs)}

\item{D}{Nominal diameter (ft) of runner}

\item{rpm}{Runner revolutions per minute}

\item{opt}{Ratio of turbine discharge at best efficiency to hydraulic capacity}

\item{radius_ratio}{r/R where R = 0.5 * D; passage near hub (0.5), mid-blade (0.75), blade tip (1)}
}
\description{
Relative flow angle (rad) at turbine discharge; used in strike equations
}
